+++ omni/ComfyUI-OmniXPU/fixes/legacy_median.py
@@ -2,7 +2,7 @@ import logging
 
 import torch
 
-from .debug import log_debug_event, trace_patch
+from ..patches.debug import log_debug_event, trace_patch
 
 log = logging.getLogger("ComfyUI-OmniXPU")